+++ R/gg_hpointrange.R
@@ -1,7 +1,7 @@
 #' @title Horizontal pointrange ggplot.
 #' @description Horizontal pointrange ggplot that is not coloured and not facetted.
 #' @param data An ungrouped summarised tibble or dataframe in a structure to be plotted untransformed. Required input.
-#' @param x_var Unquoted numeric variable to be on the x scale. Required input.
+#' @param xmiddle_var Unquoted numeric variable to be on the x scale. Required input.
 #' @param xmin_var Unquoted numeric variable to be the minimum of the x vertical line. Required input.
 #' @param xmax_var Unquoted numeric variable to be the maximum of the x vertical line. Required input.
 #' @param y_var Unquoted variable to be on the y scale (i.e. character, factor, logical, numeric, date or datetime). If numeric, date or datetime, variable values are bins that are mutually exclusive and equidistant. Required input.
@@ -19,11 +19,11 @@
 #' @param x_breaks_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 5.
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or named vector to modify x scale labels. Use function(x) x to keep labels untransformed.
-#' @param x_na_rm TRUE or FALSE of whether to include x_var NA values. Defaults to FALSE.
+#' @param x_na_rm TRUE or FALSE of whether to include xmiddle_var NA values. Defaults to FALSE.
 #' @param x_title X scale title string. Defaults to "".
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
 #' @param x_zero For a numeric x variable, TRUE or FALSE of whether the minimum of the x scale is zero. Defaults to FALSE.
-#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in x_var. Otherwise defaults to FALSE.  
+#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in xmiddle_var. Otherwise defaults to FALSE.  
 #' @param y_balance For a numeric y variable, add balance to the y scale so that zero is in the centre. Defaults to FALSE.
 #' @param y_breaks_n For a numeric or date y variable, the desired number of intervals on the y scale, as calculated by the pretty algorithm. Defaults to 5. 
 #' @param y_expand A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
@@ -54,7 +54,7 @@
 #' 
 #' gg_hpointrange(
 #'   plot_data,
-#'   x_var = middle,
+#'   xmiddle_var = middle,
 #'   xmin_var = lower,
 #'   xmax_var = upper,
 #'   y_var = sex,
@@ -62,7 +62,7 @@
 #'   y_na_rm = TRUE)
 #'   
 gg_hpointrange <- function(data,
-                    x_var,
+                    xmiddle_var,
                     xmin_var,
                     xmax_var,
                     y_var,
@@ -104,7 +104,7 @@ gg_hpointrange <- function(data,
   data <- dplyr::ungroup(data)
   
   #quote
-  x_var <- rlang::enquo(x_var) #numeric var
+  xmiddle_var <- rlang::enquo(xmiddle_var) #numeric var
   xmin_var <- rlang::enquo(xmin_var) #numeric var
   xmax_var <- rlang::enquo(xmax_var) #numeric var
   y_var <- rlang::enquo(y_var)
@@ -113,7 +113,7 @@ gg_hpointrange <- function(data,
   #na's
   if (x_na_rm == TRUE) {
     data <- data %>% 
-      dplyr::filter(!is.na(!!x_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
+      dplyr::filter(!is.na(!!xmiddle_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
   }
   if (y_na_rm == TRUE) {
     data <- data %>% 
@@ -121,11 +121,11 @@ gg_hpointrange <- function(data,
   }
   
   #vectors
-  x_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
+  xmiddle_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
   y_var_vctr <- dplyr::pull(data, !!y_var)
 
   #warnings
-  if (!is.numeric(x_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
+  if (!is.numeric(xmiddle_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
   
   #logical to factor
   if (is.logical(y_var_vctr)) {
@@ -136,7 +136,7 @@ gg_hpointrange <- function(data,
   }
   
   #titles sentence case
-  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
+  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(xmiddle_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
   
   #reverse & reorder
@@ -159,7 +159,7 @@ gg_hpointrange <- function(data,
     coord_flip(clip = "off") +
     theme +
     geom_linerange(aes(x = !!y_var, ymin = !!xmin_var, ymax = !!xmax_var, text = !!text_var), size = size_line, col = pal_line) +
-    geom_point(aes(x = !!y_var, y = !!x_var, text = !!text_var), col = pal_point, alpha = alpha_point, size = size_point) 
+    geom_point(aes(x = !!y_var, y = !!xmiddle_var, text = !!text_var), col = pal_point, alpha = alpha_point, size = size_point) 
   
   #y scale 
   if (is.numeric(y_var_vctr) | lubridate::is.Date(y_var_vctr) | lubridate::is.POSIXt(y_var_vctr)) {
@@ -210,16 +210,16 @@ gg_hpointrange <- function(data,
   }
   
   #x scale
-  x_zero_list <- sv_x_zero_adjust(x_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
+  x_zero_list <- sv_x_zero_adjust(xmiddle_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
   x_zero <- x_zero_list[[1]]
   x_zero_line <- x_zero_list[[2]]
   
-  if (all(x_var_vctr == 0, na.rm = TRUE)) {
+  if (all(xmiddle_var_vctr == 0, na.rm = TRUE)) {
     plot <- plot +
       scale_y_continuous(expand = x_expand, breaks = c(0, 1), labels = x_labels, limits = c(0, 1))
   }
   else ({
-    x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = mobile)
+    x_breaks <- sv_numeric_breaks_h(xmiddle_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     
     plot <- plot +
@@ -260,7 +260,7 @@ gg_hpointrange <- function(data,
 #' @title Horizontal pointrange ggplot that is coloured.
 #' @description Horizontal pointrange ggplot that is coloured, but not facetted.
 #' @param data An ungrouped summarised tibble or dataframe in a structure to be plotted untransformed. Required input.
-#' @param x_var Unquoted numeric variable to be on the x scale. Required input.
+#' @param xmiddle_var Unquoted numeric variable to be on the x scale. Required input.
 #' @param xmin_var Unquoted numeric variable to be the minimum of the x vertical line. Required input.
 #' @param xmax_var Unquoted numeric variable to be the maximum of the x vertical line. Required input.
 #' @param y_var Unquoted variable to be on the y scale (i.e. character, factor, logical, numeric, date or datetime). If numeric, date or datetime, variable values are bins that are mutually exclusive and equidistant. Required input.
@@ -281,11 +281,11 @@ gg_hpointrange <- function(data,
 #' @param x_breaks_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 2. 
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or named vector to modify x scale labels. Use function(x) x to keep labels untransformed.
-#' @param x_na_rm TRUE or FALSE of whether to include x_var NA values. Defaults to FALSE.
+#' @param x_na_rm TRUE or FALSE of whether to include xmiddle_var NA values. Defaults to FALSE.
 #' @param x_title X scale title string. Defaults to "".
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
 #' @param x_zero For a numeric x variable, TRUE or FALSE of whether the minimum of the x scale is zero. Defaults to FALSE.
-#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in x_var. Otherwise defaults to FALSE.  
+#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in xmiddle_var. Otherwise defaults to FALSE.  
 #' @param y_balance For a numeric y variable, add balance to the y scale so that zero is in the centre. Defaults to FALSE.
 #' @param y_breaks_n For a numeric or date y variable, the desired number of intervals on the y scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param y_dodge The amount to dodge pointranges by along the y axis. Defaults to 0 (i.e. identity).
@@ -327,7 +327,7 @@ gg_hpointrange <- function(data,
 #' 
 #' gg_hpointrange_col(
 #'   plot_data,
-#'   x_var = middle,
+#'   xmiddle_var = middle,
 #'   xmin_var = lower,
 #'   xmax_var = upper,
 #'   y_var = species,
@@ -337,7 +337,7 @@ gg_hpointrange <- function(data,
 #'   y_dodge = 0.2)
 #'   
 gg_hpointrange_col <- function(data,
-                        x_var,
+                        xmiddle_var,
                         xmin_var,
                         xmax_var,
                         y_var,
@@ -393,7 +393,7 @@ gg_hpointrange_col <- function(data,
   data <- dplyr::ungroup(data)
   
   #quote
-  x_var <- rlang::enquo(x_var) #numeric var
+  xmiddle_var <- rlang::enquo(xmiddle_var) #numeric var
   xmin_var <- rlang::enquo(xmin_var) #numeric var
   xmax_var <- rlang::enquo(xmax_var) #numeric var
   y_var <- rlang::enquo(y_var) 
@@ -403,7 +403,7 @@ gg_hpointrange_col <- function(data,
   #na's
   if (x_na_rm == TRUE) {
     data <- data %>% 
-      dplyr::filter(!is.na(!!x_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
+      dplyr::filter(!is.na(!!xmiddle_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
   }
   if (y_na_rm == TRUE) {
     data <- data %>% 
@@ -415,12 +415,12 @@ gg_hpointrange_col <- function(data,
   }
   
   #vectors
-  x_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
+  xmiddle_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
   y_var_vctr <- dplyr::pull(data, !!y_var)
   col_var_vctr <- dplyr::pull(data, !!col_var)
   
   #warnings
-  if (!is.numeric(x_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
+  if (!is.numeric(xmiddle_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
   
   if (!is.null(col_method)) {
     if (!col_method %in% c("continuous", "bin", "quantile", "category")) stop("Please use a colour method of 'continuous', 'bin', 'quantile' or 'category'")
@@ -441,7 +441,7 @@ gg_hpointrange_col <- function(data,
   }
   
   #titles sentence case
-  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
+  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(xmiddle_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
   if (is.null(col_title)) col_title <- snakecase::to_sentence_case(rlang::as_name(col_var))
   
@@ -542,7 +542,7 @@ gg_hpointrange_col <- function(data,
     coord_flip(clip = "off") +
     theme +
     geom_linerange(aes(x = !!y_var, ymin = !!xmin_var, ymax = !!xmax_var, col = !!col_var, text = !!text_var), size = size_line, position = position_dodge(width = y_dodge)) +
-    geom_point(aes(x = !!y_var, y = !!x_var, col = !!col_var, text = !!text_var), alpha = alpha_point, size = size_point, position = position_dodge(width = y_dodge))
+    geom_point(aes(x = !!y_var, y = !!xmiddle_var, col = !!col_var, text = !!text_var), alpha = alpha_point, size = size_point, position = position_dodge(width = y_dodge))
   
   #y scale 
   if (is.numeric(y_var_vctr) | lubridate::is.Date(y_var_vctr) | lubridate::is.POSIXt(y_var_vctr)) {
@@ -593,16 +593,16 @@ gg_hpointrange_col <- function(data,
   }
   
   #x scale
-  x_zero_list <- sv_x_zero_adjust(x_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
+  x_zero_list <- sv_x_zero_adjust(xmiddle_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
   x_zero <- x_zero_list[[1]]
   x_zero_line <- x_zero_list[[2]]
   
-  if (all(x_var_vctr == 0, na.rm = TRUE)) {
+  if (all(xmiddle_var_vctr == 0, na.rm = TRUE)) {
     plot <- plot +
       scale_y_continuous(expand = x_expand, breaks = c(0, 1), labels = x_labels, limits = c(0, 1))
   }
   else ({
-    x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = mobile)
+    x_breaks <- sv_numeric_breaks_h(xmiddle_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     
     plot <- plot +
@@ -669,7 +669,7 @@ gg_hpointrange_col <- function(data,
 #' @title Horizontal pointrange ggplot that is facetted.
 #' @description Horizontal pointrange ggplot that is facetted, but not coloured.
 #' @param data An ungrouped summarised tibble or dataframe in a structure to be plotted untransformed. Required input.
-#' @param x_var Unquoted numeric variable to be on the x scale. Required input.
+#' @param xmiddle_var Unquoted numeric variable to be on the x scale. Required input.
 #' @param xmin_var Unquoted numeric variable to be the minimum of the x vertical line. Required input.
 #' @param xmax_var Unquoted numeric variable to be the maximum of the x vertical line. Required input.
 #' @param y_var Unquoted variable to be on the y scale (i.e. character, factor, logical, numeric, date or datetime). If numeric, date or datetime, variable values are bins that are mutually exclusive and equidistant. Required input.
@@ -688,11 +688,11 @@ gg_hpointrange_col <- function(data,
 #' @param x_breaks_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 2. 
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or named vector to modify x scale labels. Use function(x) x to keep labels untransformed.
-#' @param x_na_rm TRUE or FALSE of whether to include x_var NA values. Defaults to FALSE.
+#' @param x_na_rm TRUE or FALSE of whether to include xmiddle_var NA values. Defaults to FALSE.
 #' @param x_title X scale title string. Defaults to "".
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
 #' @param x_zero For a numeric x variable, TRUE or FALSE of whether the minimum of the x scale is zero. Defaults to FALSE.
-#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in x_var. Otherwise defaults to FALSE.  
+#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in xmiddle_var. Otherwise defaults to FALSE.  
 #' @param y_balance For a numeric y variable, add balance to the y scale so that zero is in the centre. Defaults to FALSE.
 #' @param y_breaks_n For a numeric or date y variable, the desired number of intervals on the y scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param y_expand A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
@@ -729,7 +729,7 @@ gg_hpointrange_col <- function(data,
 #' gg_hpointrange_facet(
 #'   plot_data,
 #'   y_var = species,
-#'   x_var = middle,
+#'   xmiddle_var = middle,
 #'   xmin_var = lower,
 #'   xmax_var = upper,
 #'   facet_var = sex,
@@ -737,7 +737,7 @@ gg_hpointrange_col <- function(data,
 #'   x_title = "Body mass g")
 #'   
 gg_hpointrange_facet <- function(data,
-                          x_var,
+                          xmiddle_var,
                           xmin_var,
                           xmax_var,
                           y_var,
@@ -785,7 +785,7 @@ gg_hpointrange_facet <- function(data,
   data <- dplyr::ungroup(data)
   
   #quote
-  x_var <- rlang::enquo(x_var) #numeric var
+  xmiddle_var <- rlang::enquo(xmiddle_var) #numeric var
   xmin_var <- rlang::enquo(xmin_var) #numeric var
   xmax_var <- rlang::enquo(xmax_var) #numeric var
   y_var <- rlang::enquo(y_var) #categorical var
@@ -795,7 +795,7 @@ gg_hpointrange_facet <- function(data,
   #na's
   if (x_na_rm == TRUE) {
     data <- data %>% 
-      dplyr::filter(!is.na(!!x_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
+      dplyr::filter(!is.na(!!xmiddle_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
   }
   if (y_na_rm == TRUE) {
     data <- data %>% 
@@ -807,12 +807,12 @@ gg_hpointrange_facet <- function(data,
   }
   
   #vectors
-  x_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
+  xmiddle_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
   y_var_vctr <- dplyr::pull(data, !!y_var)
   facet_var_vctr <- dplyr::pull(data, !!facet_var)
   
   #warnings
-  if (!is.numeric(x_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
+  if (!is.numeric(xmiddle_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
   if (is.numeric(facet_var_vctr)) stop("Please use a categorical facet variable for a pointrange plot")
   
   #logical to factor
@@ -830,7 +830,7 @@ gg_hpointrange_facet <- function(data,
   }
   
   #titles sentence case
-  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
+  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(xmiddle_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
   
   #reverse
@@ -860,7 +860,7 @@ gg_hpointrange_facet <- function(data,
     coord_flip(clip = "off") +
     theme +
     geom_linerange(aes(x = !!y_var, ymin = !!xmin_var, ymax = !!xmax_var, text = !!text_var), size = size_line, col = pal_line) +
-    geom_point(aes(x = !!y_var, y = !!x_var, text = !!text_var), col = pal_point, size = size_point)
+    geom_point(aes(x = !!y_var, y = !!xmiddle_var, text = !!text_var), col = pal_point, size = size_point)
 
   #y scale
   if (is.character(y_var_vctr) | is.factor(y_var_vctr)){
@@ -906,17 +906,17 @@ gg_hpointrange_facet <- function(data,
   }
   
   #x scale
-  x_zero_list <- sv_x_zero_adjust(x_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
+  x_zero_list <- sv_x_zero_adjust(xmiddle_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
   if (facet_scales %in% c("fixed", "free_y")) x_zero <- x_zero_list[[1]]
   x_zero_line <- x_zero_list[[2]]
   
   if (facet_scales %in% c("fixed", "free_y")) {
-    if (all(x_var_vctr == 0, na.rm = TRUE)) {
+    if (all(xmiddle_var_vctr == 0, na.rm = TRUE)) {
       plot <- plot +
         scale_y_continuous(expand = x_expand, breaks = c(0, 1), labels = x_labels, limits = c(0, 1))
     }
     else ({
-      x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = FALSE)
+      x_breaks <- sv_numeric_breaks_h(xmiddle_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_breaks), max(x_breaks))
       
       plot <- plot +
@@ -950,7 +950,7 @@ gg_hpointrange_facet <- function(data,
 #' @title Horizontal pointrange ggplot that is coloured and facetted.
 #' @description Horizontal pointrange ggplot that is coloured and facetted.
 #' @param data An ungrouped summarised tibble or dataframe in a structure to be plotted untransformed. Required input.
-#' @param x_var Unquoted numeric variable to be on the x scale. Required input.
+#' @param xmiddle_var Unquoted numeric variable to be on the x scale. Required input.
 #' @param xmin_var Unquoted numeric variable to be the minimum of the x vertical line. Required input.
 #' @param xmax_var Unquoted numeric variable to be the maximum of the x vertical line. Required input.
 #' @param y_var Unquoted variable to be on the y scale (i.e. character, factor, logical, numeric, date or datetime). If numeric, date or datetime, variable values are bins that are mutually exclusive and equidistant. Required input.
@@ -972,12 +972,12 @@ gg_hpointrange_facet <- function(data,
 #' @param y_dodge The amount to dodge pointranges by along the y axis. Defaults to 0 (i.e. identity).
 #' @param x_expand A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
 #' @param x_labels A function or named vector to modify x scale labels. Use function(x) x to keep labels untransformed.
-#' @param x_na_rm TRUE or FALSE of whether to include x_var NA values. Defaults to FALSE.
+#' @param x_na_rm TRUE or FALSE of whether to include xmiddle_var NA values. Defaults to FALSE.
 #' @param x_breaks_n For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 5. 
 #' @param x_title X scale title string. Defaults to "".
 #' @param x_title_wrap Number of characters to wrap the x title to. Defaults to 50. 
 #' @param x_zero For a numeric x variable, TRUE or FALSE of whether the minimum of the x scale is zero. Defaults to FALSE.
-#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in x_var. Otherwise defaults to FALSE.  
+#' @param x_zero_line For a numeric x variable, TRUE or FALSE whether to add a zero reference line to the x scale. Defaults to TRUE if there are positive and negative values in xmiddle_var. Otherwise defaults to FALSE.  
 #' @param y_balance For a numeric y variable, add balance to the y scale so that zero is in the centre. Defaults to FALSE.
 #' @param y_dodge The amount to dodge pointranges by along the y axis. Defaults to 0 (i.e. identity).
 #' @param y_expand A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions. 
@@ -1026,7 +1026,7 @@ gg_hpointrange_facet <- function(data,
 #' gg_hpointrange_col_facet(
 #'   plot_data,
 #'   y_var = year,
-#'   x_var = middle,
+#'   xmiddle_var = middle,
 #'   xmin_var = lower,
 #'   xmax_var = upper,
 #'   col_var = sex,
@@ -1037,7 +1037,7 @@ gg_hpointrange_facet <- function(data,
 #'   y_dodge = 0.2)
 #'   
 gg_hpointrange_col_facet <- function(data,
-                              x_var,
+                              xmiddle_var,
                               xmin_var,
                               xmax_var,
                               y_var,
@@ -1099,7 +1099,7 @@ gg_hpointrange_col_facet <- function(data,
   data <- dplyr::ungroup(data)
   
   #quote
-  x_var <- rlang::enquo(x_var) #numeric var
+  xmiddle_var <- rlang::enquo(xmiddle_var) #numeric var
   xmin_var <- rlang::enquo(xmin_var) #numeric var
   xmax_var <- rlang::enquo(xmax_var) #numeric var
   y_var <- rlang::enquo(y_var) 
@@ -1110,7 +1110,7 @@ gg_hpointrange_col_facet <- function(data,
   #na's
   if (x_na_rm == TRUE) {
     data <- data %>% 
-      dplyr::filter(!is.na(!!x_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
+      dplyr::filter(!is.na(!!xmiddle_var), !is.na(!!xmin_var), !is.na(!!xmax_var))
   }
   if (y_na_rm == TRUE) {
     data <- data %>% 
@@ -1126,13 +1126,13 @@ gg_hpointrange_col_facet <- function(data,
   }
   
   #vectors
-  x_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
+  xmiddle_var_vctr <- c(dplyr::pull(data, !!xmin_var), dplyr::pull(data, !!xmax_var))
   y_var_vctr <- dplyr::pull(data, !!y_var)
   col_var_vctr <- dplyr::pull(data, !!col_var)
   facet_var_vctr <- dplyr::pull(data, !!facet_var)
   
   #warnings
-  if (!is.numeric(x_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
+  if (!is.numeric(xmiddle_var_vctr)) stop("Please use a numeric x variable for a pointrange plot")
   if (is.numeric(facet_var_vctr)) stop("Please use a categorical facet variable for a pointrange plot")
   
   if (!is.null(col_method)) {
@@ -1167,7 +1167,7 @@ gg_hpointrange_col_facet <- function(data,
   }
   
   #titles sentence case
-  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(x_var))
+  if (is.null(x_title)) x_title <- snakecase::to_sentence_case(rlang::as_name(xmiddle_var))
   if (is.null(y_title)) y_title <- snakecase::to_sentence_case(rlang::as_name(y_var))
   if (is.null(col_title)) col_title <- snakecase::to_sentence_case(rlang::as_name(col_var))
   
@@ -1268,7 +1268,7 @@ gg_hpointrange_col_facet <- function(data,
     coord_flip(clip = "off") + 
     theme +
     geom_linerange(aes(x = !!y_var, ymin = !!xmin_var, ymax = !!xmax_var, col = !!col_var, text = !!text_var), size = size_line, position = position_dodge(width = y_dodge)) +
-    geom_point(aes(x = !!y_var, y = !!x_var, col = !!col_var, text = !!text_var), alpha = alpha_point, size = size_point, position = position_dodge(width = y_dodge))
+    geom_point(aes(x = !!y_var, y = !!xmiddle_var, col = !!col_var, text = !!text_var), alpha = alpha_point, size = size_point, position = position_dodge(width = y_dodge))
   
   #y scale
   if (is.character(y_var_vctr) | is.factor(y_var_vctr)){
@@ -1314,17 +1314,17 @@ gg_hpointrange_col_facet <- function(data,
   }
   
   #x scale
-  x_zero_list <- sv_x_zero_adjust(x_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
+  x_zero_list <- sv_x_zero_adjust(xmiddle_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)
   if (facet_scales %in% c("fixed", "free_y")) x_zero <- x_zero_list[[1]]
   x_zero_line <- x_zero_list[[2]]
   
   if (facet_scales %in% c("fixed", "free_y")) {
-    if (all(x_var_vctr == 0, na.rm = TRUE)) {
+    if (all(xmiddle_var_vctr == 0, na.rm = TRUE)) {
       plot <- plot +
         scale_y_continuous(expand = x_expand, breaks = c(0, 1), labels = x_labels, limits = c(0, 1))
     }
     else ({
-      x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = FALSE)
+      x_breaks <- sv_numeric_breaks_h(xmiddle_var_vctr, balance = x_balance, breaks_n = x_breaks_n, zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_breaks), max(x_breaks))
       
       plot <- plot +